requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
jinja2==3.1.3
openai==1.40.3
google-api-python-client==2.148.0
google-auth==2.34.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0
beautifulsoup4==4.12.3
requests==2.32.3
python-dotenv==1.0.1
